Added check for file existance
This raises an exception when a user wants to embed base64 encoded image (File name too long - data:image/png;base64,iVBORw0KGgoAA...) into a page (to view in CI).

+++ lib/cucumber/formatter/gherkin_formatter_adapter.rb
@@ -83,11 +83,15 @@ module Cucumber
       end
 
       def embed(file, mime_type, label)
-        data = File.open(file, 'rb') { |f| f.read }
-        if defined?(JRUBY_VERSION)
-          data = data.to_java_bytes
+        # Only embed if file exists
+        if File.exists?(file)
+          data = File.open(file, 'rb') { |f| f.read }
+
+          if defined?(JRUBY_VERSION)
+            data = data.to_java_bytes
+          end
+          @gf.embedding(mime_type, data)
         end
-        @gf.embedding(mime_type, data)
       end
     end
   end
